% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file from which to read the data}
}
\value{
a \code{link{tbl_df}}
}
\description{
This function reads data reads data from a csv file into a tibble.
Internally the function uses \code{readr::read_csv} to read the file.
}
\details{
If \code{filename} does not exist an error is thrown.
}
\examples{
fars_read("data/accident_2015.csv.bz2")

}
