% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read accident data from files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years for which to read data}
}
\value{
a list of tibbles, each containing data for one year
}
\description{
This function reads data from files named in format accident_YYYY.csv.bz2
where YYYY denotes the year which the data represents. Argument \code{years}
specifies which years' data is read.
}
\details{
The function results in an error no file for a specified year is found or
the file contains invalid data.
}
\examples{
\dontrun{
 y <- c(2017, 2018, 2019)
 fars_read_years(y)
}

}
