% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name for certain year's data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer to be used in the file name to indicate the year}
}
\value{
a string in format accident_YYYY.csv.bz2 where YYYY is replaced by
     \code{year}
}
\description{
This is a helper function that creates "accident_2020.csv.bz2" style
file names.
}
\examples{
\dontrun{make_filename(2020)}

}
