% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello_World.R
\name{print_hello_world}
\alias{print_hello_world}
\title{A Hello World Function}
\usage{
print_hello_world()
}
\description{
This function outputs 'Hello World.' That is all.
}
\examples{
print_hello_world()
}

